+++ opentelemetry-proto/src/transform/mod.rs
@@ -11,6 +11,3 @@ pub mod logs;
 
 #[cfg(feature = "zpages")]
 pub mod tracez;
-
-#[cfg(feature = "profiles")]
-pub mod profiles;