+++ opentelemetry-sdk/src/metrics/mod.rs
@@ -73,8 +73,7 @@ pub use in_memory_exporter::{InMemoryMetricExporter, InMemoryMetricExporterBuild
 
 #[cfg(feature = "spec_unstable_metrics_views")]
 pub use aggregation::*;
-#[cfg(feature = "spec_unstable_metrics_views")]
-pub use error::{MetricError, MetricResult};
+
 #[cfg(feature = "experimental_metrics_custom_reader")]
 pub use manual_reader::*;
 pub use meter_provider::*;
@@ -84,9 +83,12 @@ pub use pipeline::Pipeline;
 
 pub use instrument::{Instrument, InstrumentKind, Stream, StreamBuilder};
 
+<<<<<<< HEAD
+=======
 #[cfg(feature = "spec_unstable_metrics_views")]
 pub use view::new_view;
 
+>>>>>>> f04e9ec6 (feat: Use builder pattern for constructing Metric Streams (#2984))
 pub use view::View;
 
 use std::hash::Hash;